Backend/requirements.txt
Flask==3.0.3
requests==2.32.3
Pillow==10.4.0